# gvb_server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gvb_server API文档",
        "title": "gvb_server API文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/settings/{name}": {
            "get": {
                "description": "获取指定类型的配置信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "系统配置查看",
                "operationId": "get-settings-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置类型名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新配置信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "更新配置信息",
                "operationId": "update-settings-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置类型名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "站点信息配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.SiteInfo"
                        }
                    },
                    {
                        "description": "邮箱配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.Email"
                        }
                    },
                    {
                        "description": "JWT 配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.Jwt"
                        }
                    },
                    {
                        "description": "七牛云配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.QiNiu"
                        }
                    },
                    {
                        "description": "QQ 配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.QQ"
                        }
                    },
                    {
                        "description": "上传配置",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/config.Upload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "config.Email": {
            "type": "object",
            "properties": {
                "default_from_email": {
                    "description": "默认发件人",
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "use_ssl": {
                    "description": "是否使用ssl",
                    "type": "boolean"
                },
                "use_tls": {
                    "description": "是否使用tls",
                    "type": "boolean"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "config.Jwt": {
            "type": "object",
            "properties": {
                "expires": {
                    "description": "过期时间",
                    "type": "integer"
                },
                "issuer": {
                    "description": "签发者",
                    "type": "string"
                },
                "secret": {
                    "description": "密钥",
                    "type": "string"
                }
            }
        },
        "config.QQ": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "redirect": {
                    "description": "登陆后的回调地址",
                    "type": "string"
                }
            }
        },
        "config.QiNiu": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "bucket": {
                    "description": "存储空间",
                    "type": "string"
                },
                "cdn": {
                    "description": "访问图片的地址的前缀",
                    "type": "string"
                },
                "enable": {
                    "description": "是否启用七牛云来缓存图片",
                    "type": "boolean"
                },
                "secret_key": {
                    "type": "string"
                },
                "size": {
                    "description": "存储的大小限制，单位为MB",
                    "type": "number"
                },
                "zone": {
                    "description": "存储的地区",
                    "type": "string"
                }
            }
        },
        "config.SiteInfo": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "bei_an": {
                    "type": "string"
                },
                "bilibili_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gitee_url": {
                    "type": "string"
                },
                "github_url": {
                    "type": "string"
                },
                "job": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qq_image": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "web": {
                    "type": "string"
                },
                "wechat_image": {
                    "type": "string"
                }
            }
        },
        "config.Upload": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "number"
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}